requirements.txt
PyOpenGL
PyOpenGL_accelerate
pyassimp
numpy
PyGLUT
setuptools